requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
asyncpg==0.27.0
psycopg2-binary==2.9.6
motor==3.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.18
alembic==1.7.4
fastapi-users==13.0.0
redis==5.0.8
fastapi-users[redis]==13.0.0
pymongo==4.1.1
fastapi-users[sqlalchemy]==13.0.0
